Employee Management System/requirements.txt
mysql-connector-python>=8.0.26
tabulate>=0.8.9 